==> prediction_correctness_experiments/1-entailment.ipynb <==
# Entailment Ratio

- PAPER: [Factuality Enhanced Language Models for Opn-Ended Text Generation](https://proceedings.neurips.cc/paper_files/paper/2022/hash/df438caa36714f69277daa92d608dd63-Abstract-Conference.html) 


import os
import sys

import pandas as pd

# Get the current working directory of the notebook
notebook_dir = os.getcwd()
# Add the parent directory to the system path
sys.path.append(os.path.join(notebook_dir, '../'))

import log_files
from data_processing import DataProcessing
from feature_extraction import RobertaFeatureExtraction

pd.set_option('max_colwidth', 800)
pd.set_option('display.max_columns', None)
pd.set_option('display.max_rows', None)

## Load Prediction and Observations Data

log_file_path = "data/prediction_logs"
predictions = True
predictions_df = log_files.read_data(notebook_dir, log_file_path, predictions)
predictions_df.head(7)

log_file_path = "data/observation_logs"
predictions = False
observations_df = log_files.read_data(notebook_dir, log_file_path, predictions)
observations_df.head(7)

col_name = 'Base Sentence'
predictions = DataProcessing.df_to_list(predictions_df, col_name)
observations = DataProcessing.df_to_list(observations_df, col_name)
len(predictions), len(observations)

## Extract Entailment with Roberta

- Pass in pair that I know entaiment exists

roberta_fe = RobertaFeatureExtraction()
entailment_df = roberta_fe.extract_entailment_features(predictions, observations)

entailment_df

## Filter Out CONTRADICTION Label

contridiction_filt = (entailment_df['Entailment Label'] != 'CONTRADICTION')
entails_df = entailment_df.loc[contridiction_filt]
entails_df

by_prediction_df = pd.pivot_table(entails_df, values='Entailment Label', index=['Prediction', 'Observation', 'Entailment Label'])
by_prediction_df

## Save Data

# by_prediction_df

# os.makedirs('../data/entailment', exist_ok=True)  
# by_prediction_df.to_excel('../data/entailment/entailment-v1.xlsx', sheet_name='PO Entainment')
