==> prediction_correctness_experiments/2-embeddings_with_cos_sim.ipynb <==
# Embeddings with Cosine Similarity

- Get the embeddings of predictions and observations with an entailment label of either neutral or entail.
- Embedding models: Spacy, SentenceTransformer, BERT
- Pass all models to sklearn cosine similarity

import os, sys

import numpy as np
import pandas as pd

from tqdm import tqdm

# Get the current working directory of the notebook
notebook_dir = os.getcwd()
# Add the parent directory to the system path
sys.path.append(os.path.join(notebook_dir, '../'))

import log_files
from metrics import Metrics
from data_processing import DataProcessing
from feature_extraction import SpacyFeatureExtraction, SentenceTransformerFeatureExtraction, BertFeatureExtraction

pd.set_option('max_colwidth', 800)
pd.set_option('display.max_columns', None)
pd.set_option('display.max_rows', None)

## Load Entailment Data

entailment_df = pd.read_csv('../data/entailment/entailment-v1.csv')  
entailment_df

## Extract Embeddings with Spacy

disable_components = [""]
# pred_spacy_fe = SpacyFeatureExtraction(predictions_df, "Base Sentence")
spacy_fe_pred = SpacyFeatureExtraction(entailment_df, 'Prediction')
spacy_embeddings_pred = spacy_fe_pred.sentence_feature_extraction()

spacy_fe_obser = SpacyFeatureExtraction(entailment_df, 'Observation')
spacy_embeddings_obser = spacy_fe_obser.sentence_feature_extraction()

spacy_cs_metrics = Metrics.get_cosine_similarity(spacy_embeddings_pred, spacy_embeddings_obser)
entailment_df['Spacy'] = spacy_cs_metrics
entailment_df

## Extract Embeddings with Sentence Transformer

st_fe_pred = SentenceTransformerFeatureExtraction(entailment_df, 'Prediction')
st_embeddings_pred = st_fe_pred.sentence_feature_extraction()

st_fe_obser = SentenceTransformerFeatureExtraction(entailment_df, 'Observation')
st_embeddings_obser = st_fe_obser.sentence_feature_extraction()

st_cs_metrics = Metrics.get_cosine_similarity(st_embeddings_pred, st_embeddings_obser)
entailment_df['SentenceTransformer'] = st_cs_metrics
entailment_df

## Extract Embeddings with Bert 

bert_fe_pred = BertFeatureExtraction(entailment_df, 'Prediction')
bert_embeddings_pred = bert_fe_pred.sentence_feature_extraction()

bert_fe_obser = BertFeatureExtraction(entailment_df, 'Observation')
bert_embeddings_obser = bert_fe_obser.sentence_feature_extraction()

bert_cs_metrics = Metrics.get_cosine_similarity(bert_embeddings_pred, bert_embeddings_obser)
entailment_df['BERT'] = bert_cs_metrics
entailment_df

## Extract Embeddings with TF x IDF

# tfidf_fe_pred = TfidfFeatureExtraction(entailment_df, 'Prediction')
# tfidf_pred_df = tfidf_fe_pred.feature_scores(max_features=300)
# tfidf_pred_df[:1]
# tfidf_embedding_pred = tfidf_pred_df.iloc[:1 , 2:].to_numpy()

# tfidf_fe_obser = TfidfFeatureExtraction(entailment_df, 'Observation')
# tfidf_obser_df = tfidf_fe_obser.feature_scores(max_features=300)
# tfidf_obser_df[:1]
# tfidf_embedding_obser = tfidf_obser_df.iloc[:1 , 2:].to_numpy()

# tfidf_cs_metrics = Metrics.get_cosine_similarity(tfidf_embedding_pred, tfidf_embedding_obser)
# entailment_df['TF x IDF'] = tfidf_cs_metrics
# entailment_df

pred_obser_features_df = pd.pivot_table(entailment_df, values='Entailment Label', index=['Prediction', 'Observation', 'Entailment Label', 'Spacy', 'SentenceTransformer', 'BERT'])
pred_obser_features_df

# os.makedirs('../data/entailment-cosine_similarity', exist_ok=True)  
# pred_obser_features_df.to_csv('../data/entailment/entailment-cosine_similarity-v1.csv')  

## Cosine Similarity (cs): Top K + Averaging

1. Top K per Model: collect top k cs scores for each embedding model
2. Top K for All Models: collect top k cs scores across all embedding models
3. Avg per Model: collect the average cs scores across all embedding models. per p:o, then per models
    - $ m_{1_{\text{cs}}} + m_{2_{\text{cs}}} + ... + m_{n_{\text{cs}}} / N $

top_k = 3

### Top K per Model

spacy_top_k = entailment_df.loc[:, ['Prediction', 'Observation', 'Entailment Label', 'Spacy']].sort_values(by='Spacy', ascending=False)[:top_k]
spacy_top_k

spacy_top_k_df = entailment_df.loc[:, ['Prediction', 'Observation', 'Entailment Label', 'Spacy']].sort_values(by='Spacy', ascending=False)[:top_k]
st_top_k_df = entailment_df.loc[:, ['Prediction', 'Observation', 'Entailment Label', 'SentenceTransformer']].sort_values(by='SentenceTransformer', ascending=False)[:top_k]
bert_top_k_df = entailment_df.loc[:, ['Prediction', 'Observation', 'Entailment Label', 'BERT']].sort_values(by='BERT', ascending=False)[:top_k]

top_k_per_model_df = DataProcessing.concat_dfs([spacy_top_k_df, st_top_k_df, bert_top_k_df])
top_k_per_model_df

os.makedirs('../data/entailment-top_k_per_model', exist_ok=True)  
top_k_per_model_df.to_csv('../data/entailment/entailment-top_k_per_model-v1.csv')  

### Top K for All Models

# Reshape the dataframe to get all model scores in a single column
cs_single_col_df = pd.melt(entailment_df, 
                      id_vars=['Prediction', 'Observation', 'Entailment Label'],
                      value_vars=['Spacy', 'SentenceTransformer', 'BERT'],
                      var_name='Model', 
                      value_name='Cosine Similarity')
cs_single_col_df

# Sort by score in descending order
sorted_cs_single_col_df = cs_single_col_df.sort_values(by='Cosine Similarity', ascending=False)
top_k_all_models_df = sorted_cs_single_col_df.head(top_k)
top_k_all_models_df

os.makedirs('../data/entailment-top_k_all_models', exist_ok=True)  
top_k_all_models_df.to_csv('../data/entailment/entailment-top_k_all_models-v1.csv')  

### Avg

# Create a new column with the average score across all three models
entailment_df['Average'] = entailment_df[['Spacy', 'SentenceTransformer', 'BERT']].mean(axis=1)
entailment_df


os.makedirs('../data/entailment-average_per_po_match', exist_ok=True)  
entailment_df.to_csv('../data/entailment/entailment-average_per_po_match-v1.csv')
