==> prediction_correctness_experiments/3-certify_with_llms.ipynb <==
# Prediction Correction

- Given a prediction and an observation, does this observation or group of observations certify this prediction?

import os
import sys

import pandas as pd

from tqdm import tqdm

# Get the current working directory of the notebook
notebook_dir = os.getcwd()
print(notebook_dir)
# Add the parent directory to the system path
sys.path.append(os.path.join(notebook_dir, '../'))

from text_generation_models import TextGenerationModelFactory

pd.set_option('max_colwidth', 800)
pd.set_option('display.max_columns', None)
pd.set_option('display.max_rows', None)

## Load Data

entailment_df = pd.read_csv('../data/entailment/entailment-cosine_similarity-v1.csv')  
entailment_df

from json import loads, dumps
entailment_dict = entailment_df.to_json(orient='records')
entailment_dict = loads(entailment_dict)
print(dumps(entailment_dict, indent=4))

## Properties of Both Prediction and Observations

prediction_properties = """A prediction <p> = (<p_s>, <p_t>, <p_d>, <p_o>), where it consists of the following four properties:

    1. <p_s>, any source entity in the domain.
        - Can be a person (with a name) or a  person such as a  reporter,  analyst,  expert,  top executive,  senior level person, etc), civilian.
        - Can only be an organization that is associated with the  prediction.
    2. <p_t>, any target entity in the  domain.
	    - Can be a person (with a name) or a  person such as a  reporter,  analyst,  expert,  top executive,  senior level person, etc).
        - Can only be an organization that is associated with the  prediction.
    3. <p_d>, date or time range when <p> is expected to come to fruition or when one should observe the <p>.
        - Forecast can range from a second to anytime in the future.
        - Answers the questions: "How far to go out from today?" or "Where to stop?".
    4. <p_o>,  prediction outcome.
        - Details relevant details such as outcome, a quantifiable metric, or slope.
"""

observation_properties = """An observation <o> = (<o_s>, <o_t>, <o_d>, <o_a>), where it consists of the following four properties:

    1. <o_s>, any source entity in the domain.
        - Can be a person (with a name) or a   person such as a   reporter,   analyst,   expert,   top executive,   senior level person, etc, civilian.
        - Can only be an organization that is associated with the   observation.
    2. <o_t>, any target entity in the   domain.
        - Can be a person (with a name) or a   person such as a   reporter,   analyst,   expert,   top executive,   senior level person, etc).
        - Can only be an organization that is associated with the   observation.
    3. <o_d>, date or time range when <o> is expected to come to fruition or when one should observe the <o>.
        - Forecast can range from a second to anytime in the future.
        - Answers the questions: "How far to go out from today?" or "Where to stop?".
    4. <o_o>, observation outcome.
        - Characteristics of a domain-specific outputs such as various quantifiable metrics relevant to the   domain.
"""

## Models to Certify

tgmf = TextGenerationModelFactory()

# Groq Cloud (https://console.groq.com/docs/overview)
gemma_29b_generation_model = tgmf.create_instance('gemma2-9b-it') 
llama_318b_instant_generation_model = tgmf.create_instance('llama-3.1-8b-instant') 
llama_3370b_versatile_generation_model = tgmf.create_instance('llama-3.3-70b-versatile')  
llama_guard_4_12b_generation_model = tgmf.create_instance('meta-llama/llama-guard-4-12b')  

models = [gemma_29b_generation_model, llama_318b_instant_generation_model, llama_3370b_versatile_generation_model, llama_guard_4_12b_generation_model]

## Certify

- [Dunamu-ml’s Submissions on AVERITEC Shared Task](https://aclanthology.org/2024.fever-1.7/)

def llm_certifier(data: list, prediction_properties: str, observation_properties: str):
    
    preds_obs_labels = []
    for record in tqdm(entailment_dict):
        prediction = record['Prediction']
        observation = record['Observation']
        entailment_label = record['Entailment Label']
        spacy_score = record['Spacy']
        sentence_transformer_score = record['SentenceTransformer']
        bert_score = record['BERT']

        

        prompt = f"We have {prediction_properties} and {observation_properties}. The main difference at the moment is a prediction is future tense and observation is past tense. Return if the observation ({observation}) certifies this prediction ({prediction}). The return label MUST be certify, not certify, or neither and nothing more (as in do not state why). Take into account entailment label ({entailment_label}) using roberta-large-mnli, embeddings (of prediction and observation) measuring cosine similarity score (between prediction and observation using sklearn.metrics.pairwise.cosine_similarity): ({spacy_score}) using Spacy, ({sentence_transformer_score}) using all-MiniLM-L6-v2, and ({bert_score}) using bert-base-uncased. The return label MUST be certify, not certify, or neither and nothing more (as in do not state why)." 
        # print(f"Prompt: {prompt}")
        # print()

        for model in models:  
            input_prompt = model.user(prompt)
            # print(input_prompt)  
            
            raw_text_llm_generation = model.chat_completion([input_prompt])
            # print(raw_text_llm_generation)

            for line in raw_text_llm_generation.split("\n"):
                # print(line)
                if line.strip():
                    pred_ob_label = (prediction, observation, line, model.__name__())
            preds_obs_labels.append(pred_ob_label)
    return preds_obs_labels

certifier_predictions = llm_certifier(entailment_dict, prediction_properties, observation_properties)
certifier_predictions

df = pd.DataFrame(certifier_predictions, columns=['Prediction Sentence', 'Observation Sentence', 'Label', 'Model'])
df
